Check for existence of widget before moving sidebars
When you move a block in the customizer to a different widget sidebar,
we check to see if the block has a widgetId and if so, assume it's a
valid widget that we can just remove from the original sidebar and add
to the destination sidebar. If it's an inner block of a widget, we
instead remove the block from the parent, create a widget from the block,
and add this to the new sidebar. To detect this we expect it to not have a
wwidgetId.

This can break when you take an existing widget and turn it into inner
blocks by transforming it into a Group. The inner block retains its
widgetId through this transformation, but the corresponding widget
instance is removed.

To cover this case, we now check both that the widgetId exists AND that
there is an existing widget with that id.